fs/reiser4: Import Reiser4 file system for Linux OS
Reiser4 :
Is not only a file system. It is a software framework for creation,
assembly and customizing file systems managing local (simple or logical) storage volumes of the Operating System.
Reiser4 is a successor of ReiserFS (which is also known as ReiserFS of version 3).
Reiser4 absorbed results of academic researches in the area of data storage,
which had been conducted since 1992 by engineers of Namesys labs in collaboration with
Moscow State University and Program Systems Institute of the Russian Academy of Sciences.
For historical reasons Reiser4 currently works only for Linux OS.
However, it can be easily ported to any operating system due to its modular infrastructure.

Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>