Kenrel: [PATCH] Apply CONFIG_PREEMPT_RT Patch Set -rt21
CONFIG_PREEMPT_RT :
This patch allows nearly all of the kernel to be preempted,
with the exception of a few very small regions of code ("raw_spinlock critical regions").
This is done by replacing most kernel spinlocks with mutexes that support priority inheritance,
as well as moving all interrupt and software interrupts to kernel threads.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>