kernel/sched: [PATCH] TurboSched V5 Patchset 2/8
Create a static key which allows to enable or disable TurboSched feature atruntime.

This key is added in order to enable the TurboSched feature only when
required. This helps in optimizing the scheduler fast-path when the
TurboSched feature is disabled.

Also provide get/put methods to keep track of the tasks using the
TurboSched feature and also refcount jitter tasks. This allows to enable
the feature on setting first task classified as jitter, similarly disable
the feature on unsetting of such last task.

Signed-off-by: Parth Shah <parth@linux.ibm.com>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>